Rename two modules
