+++ perun/check/meson.build
@@ -3,8 +3,8 @@ perun_check_dir = perun_dir / 'check'
 perun_check_files = files(
     '__init__.py',
     'factory.py',
-    'general_detection.py',
-    'nonparam_helpers.py',
+    'detection_kit.py',
+    'nonparam_kit.py',
 )
 
 py3.install_sources(
