+++ perun/check/nonparam_kit.py
@@ -11,7 +11,7 @@ import numpy as np
 from perun.postprocess.regression_analysis import data_provider
 from perun.utils import log
 from perun.utils.structs import PerformanceChange, ModelRecord
-import perun.check.general_detection as methods
+import perun.check.detection_kit as methods
 import perun.postprocess.regressogram.methods as rg_methods
 
 if TYPE_CHECKING:
